+++ test/default.integ.snapshot/github-runners-test.assets.json
@@ -239,30 +239,30 @@
         }
       }
     },
-    "561e607656b5cc9eea453e0b72cd4cbe3dea9ea755b05110ed69dbc09f170616": {
+    "05440f1bb470abc712b03eaa5b06eb17fc9f91794eb05a65218221a765e8dafd": {
       "displayName": "runners/Stolen Runners/Lambda/Code",
       "source": {
-        "path": "asset.561e607656b5cc9eea453e0b72cd4cbe3dea9ea755b05110ed69dbc09f170616.lambda",
+        "path": "asset.05440f1bb470abc712b03eaa5b06eb17fc9f91794eb05a65218221a765e8dafd.lambda",
         "packaging": "zip"
       },
       "destinations": {
-        "current_account-current_region-4be3a2ad": {
+        "current_account-current_region-2c3ade47": {
           "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "561e607656b5cc9eea453e0b72cd4cbe3dea9ea755b05110ed69dbc09f170616.zip",
+          "objectKey": "05440f1bb470abc712b03eaa5b06eb17fc9f91794eb05a65218221a765e8dafd.zip",
           "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
         }
       }
     },
-    "ed3bada86064f7bb8b02ac2b4c38bad5f6dd53a5314489a730d5905c15ebcb87": {
+    "570ec710511fe0812b4bd8f3a6bf7c92743a95a0a9bc2583074ee12040a89974": {
       "displayName": "runners/Webhook Handler/webhook-handler/Code",
       "source": {
-        "path": "asset.ed3bada86064f7bb8b02ac2b4c38bad5f6dd53a5314489a730d5905c15ebcb87.lambda",
+        "path": "asset.570ec710511fe0812b4bd8f3a6bf7c92743a95a0a9bc2583074ee12040a89974.lambda",
         "packaging": "zip"
       },
       "destinations": {
-        "current_account-current_region-6bb93fd4": {
+        "current_account-current_region-a306ae47": {
           "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "ed3bada86064f7bb8b02ac2b4c38bad5f6dd53a5314489a730d5905c15ebcb87.zip",
+          "objectKey": "570ec710511fe0812b4bd8f3a6bf7c92743a95a0a9bc2583074ee12040a89974.zip",
           "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
         }
       }
@@ -323,16 +323,16 @@
         }
       }
     },
-    "dd15ca8ab8a55e839ae3cc6ba64d77b4768d7f465e6e33061be903171b377e6f": {
+    "19ce1b34b74dbb8f4ed092914ac0d5f47958ce2baf6a141113f10aaa85810819": {
       "displayName": "github-runners-test Template",
       "source": {
         "path": "github-runners-test.template.json",
         "packaging": "file"
       },
       "destinations": {
-        "current_account-current_region-22f0d4cc": {
+        "current_account-current_region-a310ce8c": {
           "bucketName": "cdk-hnb659fds-assets-${AWS::AccountId}-${AWS::Region}",
-          "objectKey": "dd15ca8ab8a55e839ae3cc6ba64d77b4768d7f465e6e33061be903171b377e6f.json",
+          "objectKey": "19ce1b34b74dbb8f4ed092914ac0d5f47958ce2baf6a141113f10aaa85810819.json",
           "assumeRoleArn": "arn:${AWS::Partition}:iam::${AWS::AccountId}:role/cdk-hnb659fds-file-publishing-role-${AWS::AccountId}-${AWS::Region}"
         }
       }
